make CurrentProcess a final class
