+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - `Innmind\OperatingSystem\Ports` is now a final class
 - `Innmind\OperatingSystem\Filesystem` is now a final class
 - `Innmind\OperatingSystem\CurrentProcess\Signals` is now a final class
+- `Innmind\OperatingSystem\CurrentProcess` is now a final class
 
 ### Fixed
