+++ src/CurrentProcess.php
@@ -7,22 +7,61 @@ use Innmind\OperatingSystem\CurrentProcess\Signals;
 use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
 use Innmind\TimeContinuum\Period;
+use Innmind\TimeWarp\Halt;
+use Innmind\Signals\Handler;
 use Innmind\Immutable\{
     Attempt,
     SideEffect,
 };
 
-interface CurrentProcess
+final class CurrentProcess
 {
+    private Halt $halt;
+    private ?Signals $signals = null;
+
+    private function __construct(Halt $halt)
+    {
+        $this->halt = $halt;
+    }
+
+    /**
+     * @internal
+     */
+    public static function of(Halt $halt): self
+    {
+        return new self($halt);
+    }
+
     /**
      * @return Attempt<Pid>
      */
-    public function id(): Attempt;
-    public function signals(): Signals;
+    public function id(): Attempt
+    {
+        $pid = \getmypid();
+
+        /** @psalm-suppress ArgumentTypeCoercion */
+        return match ($pid) {
+            false => Attempt::error(new \RuntimeException('Failed to retrieve process id')),
+            default => Attempt::result(new Pid($pid)),
+        };
+    }
+
+    public function signals(): Signals
+    {
+        return $this->signals ??= Signals::of(new Handler);
+    }
 
     /**
      * @return Attempt<SideEffect>
      */
-    public function halt(Period $period): Attempt;
-    public function memory(): Bytes;
+    public function halt(Period $period): Attempt
+    {
+        return ($this->halt)($period);
+    }
+
+    public function memory(): Bytes
+    {
+        /** @psalm-suppress ArgumentTypeCoercion */
+        return Bytes::of(\memory_get_usage());
+    }
 }
