--- src/CurrentProcess/Generic.php
@@ -1,62 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\OperatingSystem\CurrentProcess;
-
-use Innmind\OperatingSystem\CurrentProcess;
-use Innmind\Server\Control\Server\Process\Pid;
-use Innmind\Server\Status\Server\Memory\Bytes;
-use Innmind\TimeContinuum\Period;
-use Innmind\TimeWarp\Halt;
-use Innmind\Signals\Handler;
-use Innmind\Immutable\Attempt;
-
-final class Generic implements CurrentProcess
-{
-    private Halt $halt;
-    private ?Signals $signals = null;
-
-    private function __construct(Halt $halt)
-    {
-        $this->halt = $halt;
-    }
-
-    /**
-     * @internal
-     */
-    public static function of(Halt $halt): self
-    {
-        return new self($halt);
-    }
-
-    #[\Override]
-    public function id(): Attempt
-    {
-        $pid = \getmypid();
-
-        /** @psalm-suppress ArgumentTypeCoercion */
-        return match ($pid) {
-            false => Attempt::error(new \RuntimeException('Failed to retrieve process id')),
-            default => Attempt::result(new Pid($pid)),
-        };
-    }
-
-    #[\Override]
-    public function signals(): Signals
-    {
-        return $this->signals ??= Signals::of(new Handler);
-    }
-
-    #[\Override]
-    public function halt(Period $period): Attempt
-    {
-        return ($this->halt)($period);
-    }
-
-    #[\Override]
-    public function memory(): Bytes
-    {
-        /** @psalm-suppress ArgumentTypeCoercion */
-        return Bytes::of(\memory_get_usage());
-    }
-}
